+++ src/extension/extension.ts
@@ -91,24 +91,6 @@ class Extension {
     return label;
   }
 
-  resetPermissions() {
-    this.config.browser.permissions.remove({
-      permissions: [
-        'activeTab',
-        'history',
-        'tabs',
-        'webRequest',
-        'webRequestBlocking',
-      ],
-      origins: [
-        ...this.config.permissions.sso,
-        ...this.config.permissions.console,
-        ...this.config.permissions.signin,
-        ...this.config.permissions.containers,
-      ],
-    });
-  }
-
   async checkPermissions(): Promise<ExtensionPermissions> {
     this.log('checkPermissions');
     const history = this.platform === 'safari' ? Promise.resolve(false) : this.config.browser.permissions.contains({
@@ -144,46 +126,6 @@ class Extension {
     this.log(data);
     return data;
   }
-
-  async searchHistory(): Promise<string[]> {
-    const dirs: string[] = [];
-    return this.config.browser.history.search({
-      text: 'awsapps.com/start#/',
-      startTime: (Date.now() - (1000 * 60 * 60 * 24 * 30)), // 1 month ago,
-      maxResults: 1000,
-    }).then((results) => {
-      results?.forEach((site) => {
-        const match = this.ssoUrlRegex.exec(site.url as string);
-        if (match?.groups != null) {
-          if (!(match.groups.directoryId in dirs)) {
-            dirs.push(match.groups.directoryId);
-          }
-        }
-      });
-      const uniqDirs = [...new Set(dirs)];
-      this.log(uniqDirs);
-      return uniqDirs;
-    });
-  }
-
-  checkProfiles(appProfiles: AppData[]): AppData[] {
-    this.log(appProfiles);
-    return appProfiles.map((ap) => JSON.parse(ap[Object.keys(ap)[0]]));
-  }
-
-  /*
-  static calculateChecksum(c) {
-    // generate csrf token
-    let a = 1;
-    let b = 0;
-    if (!c) { return 0; }
-    for (let i = 0; i < c.length; ++i) {
-      a = (a + c.charCodeAt(i)) % 65521;
-      b = (b + a) % 65521;
-    }
-    return (b << 15) | a;
-  }
-  */
 
   async loadIamLogins(): Promise<IamRole[]> {
     const loginsKey = `${this.config.name}-iam-logins`;
