+++ src/utils/index.ts
@@ -52,28 +52,6 @@ function getFontColor(hexcolor): 'black' | 'white' {
   return (yiq >= 180) ? 'black' : 'white';
 }
 
-// TODO cleanup after 1.5.0 release
-async function migrateData143() {
-  let users = await extension.loadUsers();
-  users = users.sort((a, b) => ((a.updatedAt > b.updatedAt) ? -1 : 1));
-  const appProfileIds = users.map((u) => u.appProfileIds);
-  const uniqProfileIds = [...new Set(appProfileIds.flat(1))];
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  const appProfiles: Array<Promise<Record<string, any>>> = [];
-  uniqProfileIds.forEach((apId) => {
-    appProfiles.push(extension.config.browser.storage.sync.get(apId));
-  });
-  Promise.all(appProfiles).then((aps) => {
-    const parsed = aps.map((ap) => JSON.parse(ap[Object.keys(ap)[0]]));
-    parsed.forEach((appProfile) => {
-      // eslint-disable-next-line vue/max-len
-      extension.saveData(appProfile.profile?.id, appProfile, extension.config.browser.storage.local);
-    });
-    const removed = extension.config.browser.storage.sync.remove(uniqProfileIds);
-    removed.then(() => { extension.log('migrateData143:removed'); });
-  });
-}
-
 export {
-  waitForElement, createFirefoxContainer, migrateData143, getFontColor,
+  waitForElement, createFirefoxContainer, getFontColor,
 };
